# swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "Askito Rest api",
      "description": "A community for developers to share their knowledge"
    },
    "host": "askito.herokuapp.com",
    "basePath": "/api/",
...
    "schemes": ["https"],
    "consumes": ["application/json"],
    "produces": ["application/json"],
    "paths": {
      "/users/register": {
        "post": {
          "tags": ["Users"],
          "summary": "Create new user in system",
          "description": "",
          "operationId": "addUser",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "description": "User object that needs to be added to the system",
              "required": true,
              "schema": {
              "$ref": "#/definitions/User"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "user created successfully"
            },
            "400": {
              "description": "User already exists (or) try different username"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/users/login":{
        "post": {
          "tags": ["Users"],
          "summary": "Login user in system",
          "description": "",
          "operationId": "LoginUser",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "description": "User object that needs to be added to the system",
              "required": true,
              "schema": {
              "$ref": "#/definitions/Login"
              }
            }
          ],
          "responses": {
            
            "200": {
              "description": "Login successfully"
            },
            "400": {
              "description": "Invalid email or password"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/users/forgot": {
        "post": {
          "tags": ["Users"],
          "summary": "sends password reset link to user email.",
          "description": "",
          "operationId": "ResetPassword",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "description": "User object that needs to be added to the system",
              "required": true,
              "schema": {
              "$ref": "#/definitions/Reset"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "password reset email sent successfully to your email"
            },
            "404": {
              "description": "user not found"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/users/reset": {
        "post": {
          "tags": ["Users"],
          "summary": "sends password reset link to user email.",
          "description": "",
          "operationId": "ResetPassword",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "password reset email sent successfully to your email"
            },
            "404": {
              "description": "user not found"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/users/reset/{token}": {
        "post": {
          "tags": ["Users"],
          "summary": "Resets password for user",
          "description": "",
          "operationId": "ResetPassword",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "token",
              "description": "token need to in params",
              "required": true,
              "type": "string"
            },
            {
              "in": "body",
              "name": "password",
              "description": "User password that needs to be updated",
              "required": true,
              "type" : "string"
            }
          ],
          "responses": {
            "200": {
              "description": "password reset successfully"
            },
            "404": {
              "description": "user not found"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/users/get/user/{name}": {
        "get": {
          "tags": ["Users"],
          "summary": "Get User profile details by name",
          "description": "",
          "operationId": "getUserProfileByName",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "name",
              "description": "name of user to be fetched",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "user fetched successfully",
              "schema": {
                "$ref": "#/definitions/User"
              }
            },
            "404": {
              "description": "user not found"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/users/get/me": {
        "get": {
          "tags": ["Users"],
          "summary": "Get User profile details of login user",
          "description": "",
          "operationId": "getUserProfile",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "user fetched successfully",
              "schema": {
                "$ref": "#/definitions/User"
              }
            },
            "404": {
              "description": "user not found"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/users/get/verify":{
        "get": {
          "tags": ["Users"],
          "summary": "sends email verification email to user email",
          "description": "",
          "operationId": "VerifyUser",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "token to be verified",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "verification email sent successfully to your email"
            },
            "404": {
              "description": "invalid token"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/users/verify/{token}":{
        "get": {
          "tags": ["Users"],
          "summary": "Verify user email",
          "description": "",
          "operationId": "VerifyUser",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "token",
              "description": "token to be verified",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "user verified successfully"
            },
            "404": {
              "description": "user not found"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/users/update": {
        "put": {
          "tags": ["Users"],
          "summary": "update profile of the current user",
          "description": "",
          "operationId": "updateUserProfile",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "description": "User object that needs to be added to the system",
              "required": true,
              "schema": {
              "$ref": "#/definitions/User"
              }
            },
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "token to be verified",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "updated successfully"
            },
            "404": {
              "description": "This user doesn't exist"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/users/edit/username": {
        "put": {
          "tags": ["Users"],
          "summary": "update username of the current user",
          "description": "",
          "operationId": "updateUsername",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "body",
              "name": "name",
              "description": "name of the user that needs to be updated.",
              "required": true
            },
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "token to be verified",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "updated successfully"
            },
            "404": {
              "description": "This user doesn't exist"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/users/edit/email": {
        "put": {
          "tags": ["Users"],
          "summary": "update email of the current user",
          "description": "",
          "operationId": "updateEmail",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "body",
              "name": "email",
              "description": "email that needs to be updated.",
              "required": true
            },
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "token to be verified",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "updated successfully"
              
            },
            "404": {
              "description": "This user doesn't exist"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/users/edit/dob": {
        "put": {
          "tags": ["Users"],
          "summary": "update date of birth of the current user",
          "description": "",
          "operationId": "updateDOB",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "body",
              "name": "dob",
              "type": "date",
              "description": "dob that needs to be updated.",
              "required": true
            },
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "token to be verified",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "updated successfully"
            },
            "404": {
              "description": "This user doesn't exist"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/users/edit/gender": {
        "put": {
          "tags": ["Users"],
          "summary": "update gender of the current user",
          "description": "",
          "operationId": "updateGender",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "body",
              "name": "gender",
              "description": "gender that needs to be updated.",
              "required": true
            },
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "token to be verified",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "updated successfully"
            },
            "404": {
              "description": "This user doesn't exist"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
      "/users/edit/expertin": {
        "put": {
          "tags": ["Users"],
          "summary": "update expertIn of the current user",
          "description": "",
          "operationId": "updateExperIn",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "body",
              "name": "expertIn",
              "description": "expertIn array that needs to be updated.",
              "required": true
            },
            {
              "in": "header",
              "name": "x-auth-token",
              "description": "token to be verified",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "updated successfully"
            },
            "404": {
              "description": "This user doesn't exist"
            },
            "500": {
              "description": "something went wrong try after some time!"
            }
          }
        }
      },
...
    },
      "definitions": {
        "User": {
          "type": "object",
          "properties": {
            "name": {
              "type": "string"
            },
            "email": {
              "type": "string"
            },
            "password": {
              "type": "string"
            },
            "dob": {
              "type":"date",
              "required":false
            },
            "gender": {
              "type":"string",
              "required":false
            },
            "expertIn": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "required":false
            } ,
            "bio": {
              "type":"string",
              "required":false
            },
            "url": {
              "type":"string",
              "required":false
            },
            "twitterUrl": {
              "type":"string",
              "required":false
            },
            "instagramUrl": {
              "type":"string",
              "required":false
            },
            "location": {
              "type":"string",
              "required":false
            }
            
          }
        },
        "Login":{
          "type": "object",
          "properties": {
            "email": {
              "type": "string"
            },
            "password": {
              "type": "string"
            }
          }
        },
        "Reset":{
          "type": "object",
          "properties": {
            "email": {
              "type": "string"
            }
          }
        },
...
      } 
}